Makes modal dropdown appear in the correct place on a scrolling page
On a longer page that scrolls, the modal dropdown (which appears when an
iscDropdown is clicked open) was appearing in the wrong place, often
offscreen. The changes to iscDropdownSelectionList.js address this.

When multiple iscDropdowns are stacked next to each other, as in a form,
there was a problem with the modal dropdown being obscured by other
non-activated iscDropdowns nearby. So pushing the z-index out for the
modal dropdown addresses this.