Parameterizes a positioning choice
When using the iscDropdown from hs-cmc, we need to set the modal's top
property to elPositionTop, or it appears offscreen when a page is
scrolled. hs-provider-directory needs the top to be elTop instead.

This change parameterizes this choice and sets the default to be elTop,
which is what PD (and possibly all other projects) need, so no code
changes should be needed there to accommodate this change. In hs-cmc,
the form wrapper directive will set useFormPositioning="true" so it will
be covered on the CMC side.