Makes dropdown lists autosize width
Wraps each dropdown item in an .isc-dropdown-item-inner span to allow
white-space:nowrap and provide enough padding to account for a
scrollbar, if one is needed. Also sets the width on the main dropdown
element to auto if it is not explicitly specified as drop-width.

The addition of the id property to the hidden input lets
angular.element('#'+scope.dropId) find the element in the controller.