Adds option to bypass typeahead filter
Normally the typeahead directive filters the results in listData based
on what is typed into the data model. In some cases, such as using the
data model to call an API for dynamic results that may not include what
is entered in the data model, this filter needs to be bypassed and all
results in the result list should show. This change enables that
behavior while preserving the existing behavior (using the filter) as
the default.