Support for asynchronous app initialization
Adds a configurable iscStateInit service and calls it from the resolve
function in the topmost 'index' state. This is needed to ensure that any
asynchronous functions that need to be run during app startup have
completed.

This service may be configured with its config function, typically
during the custom app's main module run phase.