Fixes issue with displaying validation on selects
Adding support for primitive lists to selects created an issue with
validation, where the field would be flagged as $invalid but the error
does not propagate up to the formly field scope. This seems to be
related to having two selects with the same ng-model (even if only one
is in the DOM at a time, due to ng-if). This combines the two into a
single select and moves the object/primitive switch into the ng-options
expression.