Refactors button configuration into its own object
Removes viewConfig attribute on iscForm and adds a new buttonConfig
object. The button behaviors from formConfig have been refactored into
buttonConfig so that everything for buttons is in one config object.

Documentation is in jsdocs in iscForm.js. The new config structure is
below and provides more control over the display of the default buttons
(submit and cancel), as well as the ability to include arbitrarily many
more buttons with callbacks.

Button configurations are objects with the following properties:
*     onClick    : a function to call when the button is clicked
*     afterClick : a function callback called after the button click
event has completed
*     cssClass   : css styles to apply to the button
*     text       : an i18n translation key or literal text to render on
the button
*     order      : order in which to render the button
*     hide       : if truthy, the button will not be rendered in the UI

Also refactors the api in iscFormInternal to use the same buttonConfig
object instead of submitFormApi and cancelFormApi. Note that in order to
easily support $validation, the submit button is still handled as a
type='submit' and its event is handled by iscFormInternal, not
iscSubform like other buttons.