Makes form API endpoints configurable
Adds a `getConfigUrl` function to `apiHelper` (in core tools) which
takes a config object. This allows endpoints for core module APIs to be
configurable in `app.config`.  Defaults are set in app.config in the
core tools appModule.

Refactors the form module APIs to use this configuration-based approach.

Also removes the core repo `apiHelper`, since that is provided by using
the core tools appModule to set up app-specific code.