Support for arbitrary iscForm data model
This allows the `model` attribute on an `<isc-form> ` to be set to an
arbitrary data model. If fetching a form by id, the `id` attribute may
still be used, but in cases where the id may not be available but a
model is still needed (such as when a patient requests his or her own
care plan), any object model may be used to fill in a form.