Autosave, form configuration
Adds support for autosaving a form. The frequency is configurable in the
FDN at the level of the form with the `autosave.trigger` property. The
following values are supported:

* "modelChange" - saves the form to the API after any change to the
model. Setting `forms.updateOn` to "blur" in the app configuration is
recommended if using this option, although the save call is throttled to
once/500ms.
* "pageChange" - saves the form after changing pages in a multi-page
form, if the model is dirty.
* "submit" - saves the form only when the Submit button is pressed. If
not specified, this is the default `autosave.trigger` behavior.

Makes the API call for both saving and loading the data model for a form
configurable. A `formDataApi` property has been added to the form's
`formConfig` object with the following properties:

* `wrap` - the function to wrap any call to the save api with form
metadata
* `unwrap` - the function needed to unwrap any call to the load api.
This should return the root-level data model for the form.
* `load` - the API call to load form data
* `save` - the API call to save form data

Defaults and function arguments for these properties are in the
documentation for `iscForm`. Forms using the defaults provided by
`iscForm` previously will be unaffected by this refactor.

Refactors the following form configuration options so they are
configurable with `app.config`, and puts the default values into
`isc.components.config`. The `updateOn` trigger behavior applies only to
templates not in `updateOnExcluded`.

``` javascript
'forms'          : {
'debounce'             : 75,
'allowInvalid'         : true,
'updateOn'             : 'change',
'updateOnExcluded': [
'checkbox',
'multiCheckbox',
'radio',
'select',
'dateComponents',
'dateComponentsPartial'
]
}
```

Corrects typo in `dateComponentsPartial` formly template definition.

Removes one-time binding on table cell display template so it will
update when subform data is updated.