HSPC-2971
In many version of IE, text inputs have an X icon at the right end of the input. Clicking this fires an 'input' event which does not trigger a validation check. This PR causes 'input' events in text inputs to also fire a 'change' event so that validation is updated.