Refactors customizable FDN table templates to have more semantic property syntax; also refactors the "computedTableField" widget to use existing properties instead of custom properties. So these properties: ``` javascript { 	tableHeaderLabel     : "" 	tableCellType        : "", 	tableCellTemplate    : "", 	tableCellTemplateUrl : "", 	computedField        : { 		template    : "", 		templateUrl : "" 	} }
Have become these properties:
``` javascript
{
	collections : {
		tableCell : {
			headerLabel : "",
			type        : "",
			template    : "",
			templateUrl : ""
		}
	}
}
```

This commit also adds a backwards compatibility transformation function to the `formly` transform set that maps any legacy FDN properties to their new places. This is run automatically when an FDN is loaded by the engine and outputs a deprecation warning if legacy properties are present. This will allow FDN that uses older syntax to be consumed by the latest version of the forms engine.