HSFW-158
Migrates hspcLanguageService to UIFW as iscLocalizationService. This adds a `configure` function for the app-specific functionality, which currently only supports the server API called (if any) to change languages, and a local callback function invoked after the current language is changed. Also emits a `LOCALIZATION_EVENTS.languageChanged` event after the change, for any locale-sensitive components to listen for and handle.