Support for HSPD-3114
Refactors the header for a section into a template and adds support for custom header actions. These can be specified under the section's "data.header.actions" array in the FDN.

An action has these properties:
```
{
  "icon"       : "fa-pencil",
  "name"       : "Edit",
  "className"  : "active",
  "mode"       : "view",
  "hide"       : "!formState.lib.canEditMasterRecord(section)",
  "expression" : "formState.lib.goToEditMode(section.name)"
}
```

"hide" and "expression" are expressions and may refer to form library functions like the above example.

Also fixes an unrelated bug where the "data.layout" of sections was not being applied to forms.